refactor: move the visualization query to a custom hook
This is more of a test of how we can split the code when using
dataEngine/query.
